Add TaskRun completed filter method
Add method FilterCompletedTaskRun and related test to
filter completed TaskRun.
