+++ tekton/testdata/TaskRun.Uncompleted.yaml
@@ -1,0 +1,2 @@
+apiVersion: tekton.dev/v1beta1
+kind: TaskRun
